requirements.txt
fastapi
aiofiles
spotipy
uvicorn
jinja2
deta
python-dotenv
python-multipart
feedparser
